pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "base-data-project"
version = "0.1.0"
description = "Standardized framework for data processing projects"
readme = "README.md"
authors = [{name = "António Alves", email = "antonio.alves@tlantic.com"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "sqlalchemy>=1.4.0",
    "click>=8.0.0",
    "pulp>=2.5.0",
    "numpy>=1.21.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
    "mypy>=0.900",
    "sphinx>=4.0.0",
]

[project.scripts]
base-data-project = "base_data_project.cli:main"

[tool.setuptools]
package-dir = {"" = "."}
include-package-data = true

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true